+++ src/constants/talkingkit/breathing.ts
@@ -1,0 +1,19 @@
+/**
+ * 호흡 연습 관련 상수
+ */
+
+/** 전체 애니메이션 지속 시간 (ms) */
+export const ANIMATION_DURATION = 10000;
+
+/** 들숨 단계 지속 시간 (ms) */
+export const INHALE_DURATION = 5000;
+
+/** 파란색 경로 시작 지점 (전체 경로의 20%) */
+export const BLUE_PATH_START = 0.2;
+
+/** 초기 공 위치 fallback 좌표 */
+export const INITIAL_BALL_POSITION = { x: 50, y: 200 };
+
+/** 초기 위치 설정 지연 시간 (ms) */
+export const INITIAL_POSITION_DELAY = 100;
+
